staging/lustre: fix build warnning on 32bit system
Building on 32bit system, I got warnings like bellow:
drivers/staging/lustre/lustre/llite/../include/lprocfs_status.h:666:7: note: expected ‘long unsigned int *’ but argument is of type ‘size_t *’
 char *lprocfs_find_named_value(const char *buffer, const char *name,

drivers/staging/lustre/lustre/lov/lov_io.c: In function ‘lov_io_rw_iter_init’:
include/asm-generic/div64.h:43:28: warning: comparison of distinct pointer types lacks a cast [enabled by default]
  (void)(((typeof((n)) *)0) == ((uint64_t *)0)); \

+++ drivers/staging/lustre/lustre/include/obd_lov.h
@@ -98,18 +98,18 @@ static inline __u32 lov_mds_md_stripecnt(int ea_size, __u32 lmm_magic)
   })
 #else
 # define lov_do_div64(n,base) ({					\
-	uint64_t __rem;							\
+	__u64 __rem;							\
 	if ((sizeof(base) > 4) && (((base) & 0xffffffff00000000ULL) != 0)) {  \
 		int __remainder;					      \
 		LASSERTF(!((base) & (LOV_MIN_STRIPE_SIZE - 1)), "64 bit lov " \
-			 "division %llu / %llu\n", (n), (uint64_t)(base));    \
+			 "division %llu / %llu\n", (n), (__u64)(base));    \
 		__remainder = (n) & (LOV_MIN_STRIPE_SIZE - 1);		\
 		(n) >>= LOV_MIN_STRIPE_BITS;				\
-		__rem = do_div(n, (base) >> LOV_MIN_STRIPE_BITS);	\
+		__rem = do_div(__u64(n), (base) >> LOV_MIN_STRIPE_BITS);   \
 		__rem <<= LOV_MIN_STRIPE_BITS;				\
 		__rem += __remainder;					\
 	} else {							\
-		__rem = do_div(n, base);				\
+		__rem = do_div(__u64(n), base);				\
 	}								\
 	__rem;								\
   })
